UI progress not working
Simplified the code of onPlaylistDownloaded and onPlaylistProcessed and removed the commented GA and FB tracking code.  Both routines now do not have any purpose other than to decrement their counters and dispatch the store, although dispatch is currently not working for onPlaylistProcessed due to the sync issues in the last commit.

processAllPlaylists() is now removed, and instead there are 2 for loops, 1 running after the other, for downloading and then processing the tracks.